refactor: Split keyhandler.go into multiple files
Split the large 514-line keyhandler.go file into smaller, more focused files:
- keyhandler_base.go: Base types (KeyHandler, KeyConfig)
- keyhandler_global.go: Global commands (quit, command mode, refresh, help, cancel)
- keyhandler_navigation.go: Navigation commands (up, down, go to start/end, back)
- keyhandler_compose.go: Docker Compose commands
- keyhandler_docker.go: Docker container management commands
- keyhandler_views.go: View switching commands
- keyhandler_file_browser.go: File browser commands
- keyhandler_search.go: Search and filter commands

This improves code organization and maintainability.
