+++ internal/ui/keyhandler_base.go
@@ -1,0 +1,13 @@
+package ui
+
+import tea "github.com/charmbracelet/bubbletea"
+
+// KeyHandler represents a function that handles a key press
+type KeyHandler func(msg tea.KeyMsg) (tea.Model, tea.Cmd)
+
+// KeyConfig represents a key binding configuration
+type KeyConfig struct {
+	Keys        []string
+	Description string
+	KeyHandler  KeyHandler
+}
